==> Scripts/3 Data import and descriptive analysis_2.ipynb <==
import pandas as pd
import numpy as np
import os

df = pd.read_csv(r'C:\Users\wsaha\Documents\Insta basket analysis\02 Data\Original data\orders.csv', index_col = False)

#setting path for original file folder

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

path

#Retrieving order file using path

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# removing eval_set

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# Var_list collumns

vars_list

df.head()

# importing procuscts csv
df_prod = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#top 20 rows

df_prod.head(20)

#bottom 35 rows

df_prod.tail(35)

#name of columns

df_prod.columns

#nr of rowssha

df_prod.shape

df_prod.describe()

#Max aisle ID

max_aisle_id = df_prod['aisle_id'].max()

max_aisle_id

#deprtment ID data type

department_id_dtype = df_prod['department_id'].dtype

department_id_dtype
