==> Scripts/4 Data wrangling and subsetts.ipynb <==
import pandas as pd
import numpy as py
import os

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

#importing orders
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

#importing products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#drop eval_set within data frame
df_ords.drop(columns = ['eval_set'])

#Deleting eval_set entirely from data set. Not like above which was only visual
df_ords = df_ords.drop(columns = ['eval_set'])

#USing value count to see missing values
df_ords['days_since_prior_order'].value_counts(dropna = False)

#renaming a collumn

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

#changing data type of order ID
df_ords['order_id'] = df_ords['order_id'].astype('str')

#Data type of order_id
df_ords['order_id'].dtype

# department table import
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

#transpose visual dept table

df_dep.T

#transpose table department

df_dep_t = df_dep.T

df_dep_t

#Adding index

df_dep_t.reset_index()

#Allocating new header 0 becaus python starts indexing from 0 he iloc[] function is used for selecting data in pandas and takes the index number of a row or column (or multiple rows and columns). By executing the function and assigning it to variable, you’re assigning the values within that row or column (or rows/columns) to that variable
new_header = df_dep_t.iloc[0]

new_header

 #Here, you can see some pretty funky syntax. What does that 1: mean? This is telling Python to copy everything from the 
#first row onward from the df_dep_t dataframe into a new dataframe, df_dep_t_new. The 1 represents row 1 in the dataframe. 
#Remember that since Python starts indexing from 0 rather than 1, “row 1” is actually the second row in the dataframe. 

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#Set headerrows as df heafer

df_dep_t_new.columns = new_header

new_header



df_dep_t_new

#pandas and create a data dictionary that contains meanings for the values in the "department_id" column

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

#using data di to obtain deparment id of 19

print(data_dict.get('19'))

#creting subset for snacks
df_snacks =  df_prods[df_prods['department_id']==19]

#filtering items within deartment id 19

df_prods[df_prods['department_id']==19]

df_snacks

#alternative way 1 using loc

df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

df_snacks_2

#alternative way 3 using loc and isin instead of ==

df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks_3

#2 user id dont need to be an integer

df_ords.head()

df_ords.info()

#changing data type of user ID
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords.info()

# 3 renaming a collumn

df_ords.rename(columns = {'order_time_of_day_' : 'hour_of_the _order_made'}, inplace = True)

#4 Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

df_ords['hour_of_the _order_made'].value_counts()

#Answer is 10  : 288418

##5 Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

print(data_dict.get('4'))

##6 The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

df_breakfast=df_prods[df_prods['department_id']==14]

df_breakfast

## 7 Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

df_dinner_parties=df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_parties

##8 How many rows does the last dataframe you created have?

df_dinner_parties.shape

## 9 customer with a "user_id" of “1.” Extract all the information you can about this user.

df_user_id_1 = df_ords[df_ords['user_id']==1]

df_user_id_1.head()

df_ords[df_ords['user_id']=='1']

#Assigning user id one from prods

df_user_id_1 = df_ords[df_ords['user_id']=='1']


df_user_id_1.head(100)

#Has 10 records with one missing value

df_user_id_1.shape

#11 You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?

df_user_id.describe()

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
