==> Scripts/5 Consistancy check.ipynb <==
import pandas as pd
import numpy as py
import os

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

#importing products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#importing orders wrangled from preped data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'), index_col = False)

#Creating a data frame
df_test = pd.DataFrame()

df_test['mix'] = ['a', 'b', 1, True]

df_test

#Checking for mixed type 

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

#Changing missing values to respective data type. 

df_test['mix'] = df_test['mix'].astype('str')

df_prods.isnull().sum()

## Obtaining missing values via subsetting

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

#Data without missing values

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean

df_prods_clean.shape

## Alternative way ** df_prods.dropna(subset = [‘product_name’], inplace = True)

#Seperating all duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

#removing duplicates

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

#Number of rows after cleaning duplicates
df_prods_clean_no_dups.shape

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

## Task 4.5

# 2

df_ords.describe()

## Based on descriptive statistics there are no anomalies in data

# 3 Check for mixed-type data in your df_ords dataframe.

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

# None of collumns have mixed data 

# If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

# None of collumns have mixed data 

# 5 Run check for missing values in your df_ords dataframe

df_ords.isnull().sum()

## days_since_prior_order has 206209 missing values

df_ords.shape

df_miss = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_miss

# 6.Address the missing values using an appropriate method.

###  206209 missing values means it is significant and further checking shows that every user has one missing value.
###  Because of that rather than replacing or deleting I have decided to leave it as it is

 # 7 Run a check for duplicate values in your df_ords data
    
df_ords_duplicates = df_ords[df_ords.duplicated()]

df_ords_duplicates

# Address the duplicates using an appropriate method.

##There are no duplicates

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_&_orders_checked.csv'))
