==> Scripts/6 Combining data frames.ipynb <==
import pandas as pd
import numpy as py
import os

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

#importing df_ords_prior.
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

df_ords_prior.head()

#importing orders wrangled from preped data
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared data', 'orders_wrangled.csv'), index_col = False)

df_ords.head()

df_ords.shape

df_ords_prior.shape

# Merging two data frames

df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

# sum up all the values in the “_merge” column
df_merged_large['_merge'].value_counts()

#Export in pickle format

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
