==> Scripts/7 Combining data frames.ipynb <==
import pandas as pd
import numpy as py
import os

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

# 3. Imoprting data frames ## Importing orders_products_combined

df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_combined.pkl'))

df_orders_products_combined.head()

# Importing df_prods

df_prods =  pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_&_orders_checked.csv'), index_col = False)

df_prods.head()

# 4 Check the shape of the imported dataframe

df_orders_products_combined.shape

df_prods.shape

# 5 Determine a suitable way to combine the orders_products_combined dataframe with your products data set

# Dropping unnamed columns in df_prods & df_orders_products_combined

df_prods = df_prods.drop(columns = ['Unnamed: 0'])


df_orders_products_combined = df_orders_products_combined.drop(columns = ['Unnamed: 0'])

df_orders_products_combined.head()

df_prods.head()

##Merging two data frames using product_id

df_orders_products_merged = df_orders_products_combined.merge(df_prods, on='product_id', how='inner', indicator='merge_indicator')

# . Confirm the results of the merge using the merge flag.
df_orders_products_merged.head()

df_orders_products_merged.shape

df_orders_products_merged['_merge'].value_counts()

# 7. Export the newly created dataframe as ords_prods_merge in a suitable format

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
