==> Scripts/8 Derive new variables.ipynb <==
import pandas as pd
import numpy as py
import os

path = r'C:\Users\wsaha\Documents\Insta basket analysis'

# Imoprting data frames ## Importing df_orders_products_merged

df_orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared data', 'orders_products_merged.pkl'))

df_orders_products_merged

df = df_orders_products_merged[:1000000]

#Creating price label column using if else statements

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

## Apply the function

df['price_range'] = df.apply(price_label, axis=1)

# oBtaining values in collumn

df['price_range'].value_counts(dropna = False)

#Confirming there are no max values

df['prices'].max()

## Using loc function instead of if else

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# running lock on entire data base

df_orders_products_merged.loc[df_orders_products_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_orders_products_merged.loc[(df_orders_products_merged['prices'] <= 15) & (df_orders_products_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_orders_products_merged.loc[df_orders_products_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_orders_products_merged['price_range_loc'].value_counts(dropna = False)

# Looking at busiest day of week 

df_orders_products_merged['orders_day_of_week'].value_counts(dropna = False)

#Use for loop to assigne bussiest and no bussiest day to each raw 

result = []

for value in df_orders_products_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# allocateing a collumn in data base for result

df_orders_products_merged ['Bussiest_day'] = result

df_orders_products_merged ['Bussiest_day'].value_counts(dropna = False)

## Task 4.7 2 Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural).
## Create a new column for this using a suitable method.

# Looking at busiest day of week 

df_orders_products_merged['orders_day_of_week'].value_counts(dropna = False)


## Two busiest days 

df_orders_products_merged.loc[df_orders_products_merged['orders_day_of_week'].isin([0, 1]), 'busiest_days'] = 'Busiest_days'

## Normal  days

df_orders_products_merged.loc[df_orders_products_merged['orders_day_of_week'].isin([6, 2, 5]), 'busiest_days'] = 'Regular_days'

## two slowest days

df_orders_products_merged.loc[df_orders_products_merged['orders_day_of_week'].isin([3, 4]), 'busiest_days'] = 'Slowest_Days'

## 3 Check the values of this new column for accuracy. Note any observations in markdown format.

df_orders_products_merged['busiest_days'].value_counts(dropna = False)

df_orders_products_merged.head()

df_orders_products_merged['busiest_days'].shape

## 4 Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” 
## Create a new column containing these labels called “busiest_period_of_day.”

# Sorting hours by value count

hours = df_orders_products_merged['hour_of_the _order_made'].value_counts(dropna = False)


hours

# create empty list and for loop to generate list of “busiest_period_of_day.”

busiest_hours = []

for row in df_orders_products_merged['hour_of_the _order_made']:
    if row in hours[:8]:
        busiest_hours.append("Most_orders")
    elif row in hours[8:16]:
        busiest_hours.append("Average_orders")
    else:
        busiest_hours.append("Fewest_orders")

# replacing the name of  busiest hour 

df_orders_products_merged['busiest_period_of_day'] = busiest_hours

#result

df_orders_products_merged.head()

## 5. Print the frequency for this new column.

df_orders_products_merged['busiest_period_of_day'].value_counts(dropna = False)

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'df_orders_products_merged_altered_updated.pkl'))
